+++ example/cmd/server/main.go
@@ -8,9 +8,9 @@ import (
 	"github.com/fluffy-bunny/grpcdotnetgo/example/internal"
 	pb "github.com/fluffy-bunny/grpcdotnetgo/example/internal/grpcContracts/helloworld"
 	_ "github.com/fluffy-bunny/grpcdotnetgo/example/internal/plugin"
-	middleware_grpc_auth "github.com/fluffy-bunny/grpcdotnetgo/middleware/auth"
-	grpcDIProtoError "github.com/fluffy-bunny/grpcdotnetgo/proto/error"
-	runtime "github.com/fluffy-bunny/grpcdotnetgo/runtime"
+	middleware_grpc_auth "github.com/fluffy-bunny/grpcdotnetgo/pkg/middleware/auth"
+	grpcDIProtoError "github.com/fluffy-bunny/grpcdotnetgo/pkg/proto/error"
+	runtime "github.com/fluffy-bunny/grpcdotnetgo/pkg/runtime"
 	pkg "github.com/fluffy-bunny/protoc-gen-go-di/pkg"
 	"github.com/gogo/protobuf/gogoproto"
 	_ "github.com/jnewmano/grpc-json-proxy/codec"
