+++ example/internal/shared.go
@@ -3,7 +3,7 @@ package internal
 import (
 	"encoding/json"
 
-	_ "github.com/fluffy-bunny/grpcdotnetgo" // ensure that go mod vendor brings everything
+	_ "github.com/fluffy-bunny/grpcdotnetgo/pkg" // ensure that go mod vendor brings everything
 )
 
 func PrettyJSON(obj interface{}) string {
