+++ example/internal/startup/startup.go
@@ -5,8 +5,6 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/fluffy-bunny/grpcdotnetgo/auth/oauth2"
-	grpcdotnetgo_core_types "github.com/fluffy-bunny/grpcdotnetgo/core/types"
 	"github.com/fluffy-bunny/grpcdotnetgo/example/internal"
 	pb "github.com/fluffy-bunny/grpcdotnetgo/example/internal/grpcContracts/helloworld"
 	backgroundCounterService "github.com/fluffy-bunny/grpcdotnetgo/example/internal/services/background/cron/counter"
@@ -14,12 +12,14 @@ import (
 	handlerGreeterService "github.com/fluffy-bunny/grpcdotnetgo/example/internal/services/helloworld/handler"
 	singletonService "github.com/fluffy-bunny/grpcdotnetgo/example/internal/services/singleton"
 	transientService "github.com/fluffy-bunny/grpcdotnetgo/example/internal/services/transient"
-	middleware_dicontext "github.com/fluffy-bunny/grpcdotnetgo/middleware/dicontext/middleware"
-	middleware_logger "github.com/fluffy-bunny/grpcdotnetgo/middleware/logger"
-	middleware_oidc "github.com/fluffy-bunny/grpcdotnetgo/middleware/oidc"
-	middleware_grpc_recovery "github.com/fluffy-bunny/grpcdotnetgo/middleware/recovery"
-	grpcDIProtoError "github.com/fluffy-bunny/grpcdotnetgo/proto/error"
-	mockoidcservice "github.com/fluffy-bunny/grpcdotnetgo/services/test/mockoidcservice"
+	"github.com/fluffy-bunny/grpcdotnetgo/pkg/auth/oauth2"
+	grpcdotnetgo_core_types "github.com/fluffy-bunny/grpcdotnetgo/pkg/core/types"
+	middleware_dicontext "github.com/fluffy-bunny/grpcdotnetgo/pkg/middleware/dicontext/middleware"
+	middleware_logger "github.com/fluffy-bunny/grpcdotnetgo/pkg/middleware/logger"
+	middleware_oidc "github.com/fluffy-bunny/grpcdotnetgo/pkg/middleware/oidc"
+	middleware_grpc_recovery "github.com/fluffy-bunny/grpcdotnetgo/pkg/middleware/recovery"
+	grpcDIProtoError "github.com/fluffy-bunny/grpcdotnetgo/pkg/proto/error"
+	mockoidcservice "github.com/fluffy-bunny/grpcdotnetgo/pkg/services/test/mockoidcservice"
 	di "github.com/fluffy-bunny/sarulabsdi"
 	grpc_ctxtags "github.com/grpc-ecosystem/go-grpc-middleware/tags"
 	_ "github.com/jnewmano/grpc-json-proxy/codec"
