+++ pkg/auth/oauth2/discovery_document.go
